pyproject.toml
[project]
name = "retail-company-langgraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "db-dtypes>=1.5.0",
    "google-cloud-bigquery>=3.40.0",
    "langchain>=1.2.9",
    "langchain-core>=1.2.8",
    "langchain-google-genai>=4.2.0",
    "langgraph>=1.0.7",
    "pandas>=2.3.3",
    "pytest>=9.0.2",
    "pytest-dotenv>=0.5.2",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "rich>=14.3.2",
    "tabulate>=0.9.0",
]
